do some package shenanigans
